feat: fallback security card
