+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 -   Security Card
+-   Fallback Security Card
 
 ### Changed
 
@@ -17,6 +18,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   The Security Card can be bound to other (currently online) players via its GUI.
 -   The binding of a Security Card can now be cleared.
 -   The Security Card tooltip and GUI now show whether the permission has been touched/changed in any way.
+-   A global (fallback) permission set for a network can be defined using the Fallback Security Card instead of using an "unbound" Security Card.
 
 ### Fixed
