+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/content/ContentIds.java
@@ -55,6 +55,7 @@ public final class ContentIds {
     public static final ResourceLocation PORTABLE_GRID = createIdentifier("portable_grid");
     public static final ResourceLocation CREATIVE_PORTABLE_GRID = createIdentifier("creative_portable_grid");
     public static final ResourceLocation SECURITY_CARD = createIdentifier("security_card");
+    public static final ResourceLocation FALLBACK_SECURITY_CARD = createIdentifier("fallback_security_card");
 
     private ContentIds() {
     }
