+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/content/ContentNames.java
@@ -27,6 +27,7 @@ public final class ContentNames {
     public static final MutableComponent NETWORK_TRANSMITTER = name("network_transmitter");
     public static final MutableComponent PORTABLE_GRID = name("portable_grid");
     public static final MutableComponent SECURITY_CARD = createTranslation("item", "security_card");
+    public static final MutableComponent FALLBACK_SECURITY_CARD = createTranslation("item", "fallback_security_card");
 
     private ContentNames() {
     }
