+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/content/CreativeModeTabItems.java
@@ -104,5 +104,6 @@ public final class CreativeModeTabItems {
         itemConsumer.accept(Items.INSTANCE.getConfigurationCard());
         itemConsumer.accept(Items.INSTANCE.getNetworkCard());
         itemConsumer.accept(Items.INSTANCE.getSecurityCard());
+        itemConsumer.accept(Items.INSTANCE.getFallbackSecurityCard());
     }
 }
