+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/content/Menus.java
@@ -11,6 +11,7 @@ import com.refinedmods.refinedstorage2.platform.common.grid.WirelessGridContaine
 import com.refinedmods.refinedstorage2.platform.common.iface.InterfaceContainerMenu;
 import com.refinedmods.refinedstorage2.platform.common.importer.ImporterContainerMenu;
 import com.refinedmods.refinedstorage2.platform.common.networking.NetworkTransmitterContainerMenu;
+import com.refinedmods.refinedstorage2.platform.common.security.FallbackSecurityCardContainerMenu;
 import com.refinedmods.refinedstorage2.platform.common.security.SecurityCardContainerMenu;
 import com.refinedmods.refinedstorage2.platform.common.storage.diskdrive.DiskDriveContainerMenu;
 import com.refinedmods.refinedstorage2.platform.common.storage.externalstorage.ExternalStorageContainerMenu;
@@ -74,6 +75,8 @@ public final class Menus {
     private Supplier<MenuType<PortableGridItemContainerMenu>> portableGridItem;
     @Nullable
     private Supplier<MenuType<SecurityCardContainerMenu>> securityCard;
+    @Nullable
+    private Supplier<MenuType<FallbackSecurityCardContainerMenu>> fallbackSecurityCard;
 
     private Menus() {
     }
@@ -244,5 +247,15 @@ public final class Menus {
 
     public void setSecurityCard(final Supplier<MenuType<SecurityCardContainerMenu>> securityCard) {
         this.securityCard = securityCard;
+    }
+
+    public MenuType<FallbackSecurityCardContainerMenu> getFallbackSecurityCard() {
+        return requireNonNull(fallbackSecurityCard).get();
+    }
+
+    public void setFallbackSecurityCard(
+        final Supplier<MenuType<FallbackSecurityCardContainerMenu>> fallbackSecurityCard
+    ) {
+        this.fallbackSecurityCard = fallbackSecurityCard;
     }
 }
