+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/security/SecurityCardModel.java
@@ -3,20 +3,14 @@ package com.refinedmods.refinedstorage2.platform.common.security;
 import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
 import com.refinedmods.refinedstorage2.platform.api.security.PlatformPermission;
 
-import java.util.UUID;
-import javax.annotation.Nullable;
-
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.item.ItemStack;
 
 class SecurityCardModel {
-    private static final String TAG_BOUND_PLAYER_ID = "bid";
-    private static final String TAG_BOUND_PLAYER_NAME = "bname";
     private static final String TAG_PERMISSIONS = "permissions";
 
-    private final ItemStack stack;
+    protected final ItemStack stack;
 
     SecurityCardModel(final ItemStack stack) {
         this.stack = stack;
@@ -55,35 +49,15 @@ class SecurityCardModel {
         permissionsTag.remove(permissionId.toString());
     }
 
-    @Nullable
-    UUID getBoundPlayerId() {
-        return (stack.getTag() == null || !stack.getTag().contains(TAG_BOUND_PLAYER_ID))
-            ? null
-            : stack.getTag().getUUID(TAG_BOUND_PLAYER_ID);
+    boolean isActive() {
+        return stack.getTag() != null && stack.getTag().contains(TAG_PERMISSIONS);
     }
 
-    @Nullable
-    String getBoundPlayerName() {
-        return (stack.getTag() == null || !stack.getTag().contains(TAG_BOUND_PLAYER_NAME))
-            ? null
-            : stack.getTag().getString(TAG_BOUND_PLAYER_NAME);
+    boolean isCleared() {
+        return stack.getTag() == null;
     }
 
-    void setBoundPlayer(@Nullable final ServerPlayer player) {
-        final CompoundTag tag = stack.getOrCreateTag();
-        if (player == null) {
-            tag.remove(TAG_BOUND_PLAYER_ID);
-            tag.remove(TAG_BOUND_PLAYER_NAME);
-            return;
-        }
-        tag.putUUID(TAG_BOUND_PLAYER_ID, player.getGameProfile().getId());
-        tag.putString(TAG_BOUND_PLAYER_NAME, player.getGameProfile().getName());
-    }
-
-    static boolean isActive(final ItemStack stack) {
-        return stack.getTag() != null
-            && stack.getTag().contains(TAG_BOUND_PLAYER_ID)
-            && stack.getTag().contains(TAG_BOUND_PLAYER_NAME)
-            && stack.getTag().contains(TAG_PERMISSIONS);
+    public void clear() {
+        stack.setTag(null);
     }
 }
