+++ refinedstorage2-platform-common/src/main/resources/assets/refinedstorage2/lang/en_us.json
@@ -225,9 +225,11 @@
   "item.refinedstorage2.security_card": "Security Card",
   "item.refinedstorage2.security_card.cleared": "Cleared binding.",
   "item.refinedstorage2.security_card.unbound": "Unbound.",
-  "item.refinedstorage2.security_card.unbound.help": "Use while crouching to bind to the current player. Right click to configure.",
+  "item.refinedstorage2.security_card.unbound.help": "Right click to configure. Use while crouching to bind to the current player.",
   "item.refinedstorage2.security_card.bound": "Bound to %s.",
-  "item.refinedstorage2.security_card.bound.help": "Use while crouching to clear binding. Right click to configure.",
+  "item.refinedstorage2.security_card.bound.help": "Right click to configure. Use while crouching to clear configuration and binding.",
+  "item.refinedstorage2.fallback_security_card": "Fallback Security Card",
+  "item.refinedstorage2.fallback_security_card.help": "Acts as a fallback if no matching Security Card for a player is found. Right click to configure. Use while crouching to clear configuration.",
   "misc.refinedstorage2.stored": "Stored: %s",
   "misc.refinedstorage2.stored_with_capacity": "Stored: %s / %s (%d%%)",
   "misc.refinedstorage2.total": "%d total",
