+++ refinedstorage2-platform-common/src/main/resources/assets/refinedstorage2/models/item/fallback_security_card.json
@@ -1,0 +1,6 @@
+{
+  "parent": "item/generated",
+  "textures": {
+    "layer0": "refinedstorage2:item/security_card/fallback"
+  }
+}
