+++ refinedstorage2-platform-common/src/main/resources/data/refinedstorage2/recipes/fallback_security_card.json
@@ -1,0 +1,14 @@
+{
+  "type": "minecraft:crafting_shapeless",
+  "ingredients": [
+    {
+      "item": "refinedstorage2:security_card"
+    },
+    {
+      "item": "minecraft:paper"
+    }
+  ],
+  "result": {
+    "item": "refinedstorage2:fallback_security_card"
+  }
+}
