+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/packet/c2s/SecurityCardPermissionPacket.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage2.platform.fabric.packet.c2s;
 
-import com.refinedmods.refinedstorage2.platform.common.security.SecurityCardContainerMenu;
+import com.refinedmods.refinedstorage2.platform.common.security.AbstractSecurityCardContainerMenu;
 
 import net.fabricmc.fabric.api.networking.v1.PacketSender;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
@@ -19,7 +19,7 @@ public class SecurityCardPermissionPacket implements ServerPlayNetworking.PlayCh
                         final PacketSender responseSender) {
         final ResourceLocation permissionId = buf.readResourceLocation();
         final boolean allowed = buf.readBoolean();
-        if (player.containerMenu instanceof SecurityCardContainerMenu securityCardContainerMenu) {
+        if (player.containerMenu instanceof AbstractSecurityCardContainerMenu securityCardContainerMenu) {
             server.execute(() -> securityCardContainerMenu.setPermission(permissionId, allowed));
         }
     }
